Update the Get Started document to reflect changes in the GitHub web UI
GitHub have shifted a few things around, renamed some labels and controls and they've also introduced a new PAT type and the original PATs are now called legacy. They differ in how permissions (scopes) for the tokens are handled. I've updated the document to describe the process for the modern PATs.

If there's interest I can edit the section to highlight how to do it for both Legacy and modern PATs. I am not doing it right away though as it is probably better to guide users to use the new PATs anyway.